make panto cursor offset relative to finger pos
instead of finger pos determining position of top of "pole", now finger
pos + spread determines cursor 3D pos

+++ Settings.h
@@ -21,7 +21,7 @@ public:
 
 	bool modeSwitched;
 
-	int positioningPointLocation[3];
+	float positioningPointLocation[3];
 	float positioningModelCoords[4]; //x,y,maxDepth,actualDepth,selectedDepth
 	float currentlySelectedPoint[3];//x,y,depth
